pyproject.toml
[build-system]
requires = ["setuptools>=80", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "sievio"
dynamic = ["version"]
description = "Repository → JSONL converter with robust decoding, structure-aware chunking, Markdown/KQL extraction, license detection, and GitHub/PDF/EVTX helpers."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Sievio Maintainers" },
]
license = { text = "MIT" }
keywords = ["jsonl", "github", "markdown", "kql", "ingestion", "chunking", "pdf", "evtx", "parquet"]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = []

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-cov>=5",
  "ruff>=0.6",
  "mypy>=1.10",
  "build>=1.2",
  "twine>=5.1",
]
tok = ["tiktoken>=0.6"]
pdf = ["pypdf>=4"]
evtx = ["python-evtx>=0.8"]
qc = [
  "torch>=2.2",
  "transformers>=4.44",
  "tiktoken>=0.6",
  "pyyaml>=6",
]
parquet = ["pyarrow>=15"]
langid = [
  "lingua-language-detector>=1.3",
  "pygments>=2.17",
]

[project.urls]
Homepage = "https://github.com/jochiraider/sievio"
Repository = "https://github.com/jochiraider/sievio"
Issues = "https://github.com/jochiraider/sievio/issues"

[project.scripts]
sievio = "sievio.cli.main:main"
sievio-qc = "sievio.cli.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["sievio*"]

[tool.setuptools_scm]
# Rely on Git tags for the version; no extra config needed for basic use.

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
ignore_missing_imports = true
